chore: 도메인 구조에 알맞은 패키지 위치 변경
- `client`폴더를 `global`폴더로 이동
- `domain`폴더를 생성하여 `chat`,`usage`,`survey` 폴더들 이동
