+++ chatbot/backend/src/main/java/com/hallachatbot/backend/domain/usage/repository/MonthlyLlmUsageRepository.java
@@ -1,0 +1,11 @@
+package com.hallachatbot.backend.domain.usage.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import com.hallachatbot.backend.domain.usage.entity.MonthlyLlmUsage;
+
+@Repository
+public interface MonthlyLlmUsageRepository extends CrudRepository<MonthlyLlmUsage, String> {
+
+}
