--- chatbot/backend/src/main/java/com/hallachatbot/backend/usage/repository/MonthlyLlmUsageRepository.java
@@ -1,11 +1,0 @@
-package com.hallachatbot.backend.usage.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import com.hallachatbot.backend.usage.entity.MonthlyLlmUsage;
-
-@Repository
-public interface MonthlyLlmUsageRepository extends CrudRepository<MonthlyLlmUsage, String> {
-
-}
